+++ src/pythonwrench/checksum.py
@@ -14,12 +14,15 @@ from types import FunctionType, MethodType
 from typing import (
     Any,
     Callable,
+    Dict,
     Generator,
     Iterable,
     Mapping,
     Optional,
     TypeVar,
     Union,
+    get_args,
+    get_origin,
     overload,
 )
 
@@ -28,8 +31,12 @@ from pythonwrench.functools import function_alias
 from pythonwrench.inspect import get_fullname
 from pythonwrench.typing import (
     DataclassInstance,
+    EllipsisType,
     NamedTupleInstance,
     NoneType,
+    is_collection_alias,
+    is_parameterized,
+    is_special_form,
 )
 
 T = TypeVar("T")
@@ -132,9 +139,7 @@ def checksum_int(x: int, **kwargs) -> int:
 # Intermediate functions
 @register_checksum_fn(bytearray)
 def checksum_bytearray(x: bytearray, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return _checksum_bytes_bytearray(x, **kwargs)
 
 
@@ -153,25 +158,25 @@ def checksum_complex(x: complex, **kwargs) -> int:
 
 @register_checksum_fn(FunctionType)
 def checksum_function(x: FunctionType, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_str(x.__qualname__, **kwargs)
 
 
 @register_checksum_fn(NoneType)
 def checksum_none(x: None, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
-    return checksum_type(x.__class__, **kwargs) + kwargs.get("accumulator", 0)
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
+    return checksum_type(x.__class__, **kwargs)
+
+
+@register_checksum_fn(EllipsisType)
+def checksum_ellipsis(x: None, **kwargs) -> int:
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
+    return checksum_type(x.__class__, **kwargs)
 
 
 @register_checksum_fn(str)
 def checksum_str(x: str, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_bytes(x.encode(), **kwargs)
 
 
@@ -183,9 +188,7 @@ def checksum_type(x: type, **kwargs) -> int:
 # Recursive functions
 @register_checksum_fn(DataclassInstance)
 def checksum_dataclass(x: DataclassInstance, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_dict(asdict(x), **kwargs)
 
 
@@ -240,9 +243,7 @@ def checksum_set(x: Union[set, frozenset], **kwargs) -> int:
 
 @register_checksum_fn(range)
 def checksum_range(x: range, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return _checksum_iterable([x.start, x.stop, x.step], **kwargs)
 
 
@@ -264,46 +265,63 @@ def checksum_method(x: MethodType, **kwargs) -> int:
 
 @register_checksum_fn(NamedTupleInstance)
 def checksum_namedtuple(x: NamedTupleInstance, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_dict(x._asdict(), **kwargs)
 
 
 @register_checksum_fn(functools.partial)
 def checksum_partial(x: functools.partial, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_list_tuple((x.func, x.args, x.keywords), **kwargs)
 
 
 @register_checksum_fn(re.Pattern)
 def checksum_pattern(x: re.Pattern, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_str(str(x), **kwargs)
 
 
 @register_checksum_fn(Path)
 def checksum_path(x: Path, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
+
     resolve_path = kwargs.get("resolve_path", False)
     if isinstance(resolve_path, bool) and resolve_path:
         x = x.expanduser().resolve()
-
     return checksum_str(str(x), **kwargs)
 
 
 @register_checksum_fn(slice)
 def checksum_slice(x: slice, **kwargs) -> int:
-    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
-        get_fullname(x)
-    )
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
     return checksum_list_tuple((x.start, x.stop, x.step), **kwargs)
+
+
+@register_checksum_fn(custom_predicate=is_parameterized)
+def checksum_parametrized(x: Any, **kwargs) -> int:
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
+    return checksum_list_tuple((get_origin(x),) + get_args(x), **kwargs)
+
+
+@register_checksum_fn(custom_predicate=is_collection_alias)
+def checksum_collection_alias(x: Any, **kwargs) -> int:
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
+    return checksum_str(x._name, **kwargs)
+
+
+@register_checksum_fn(custom_predicate=is_special_form)
+def checksum_special_form(x: Any, **kwargs) -> int:
+    kwargs = _add_type_checksum_to_accumulator(x, kwargs)
+
+    if hasattr(x, "_name"):
+        name = x._name
+    elif hasattr(x, "__name__"):
+        name = x.__name__
+    else:
+        msg = f"Unsupported argument {x=} in checksum_special_form."
+        raise ValueError(msg)
+
+    return checksum_str(name, **kwargs)
 
 
 if _CHECKSUM_PROTOCOLS:
@@ -344,7 +362,15 @@ def _checksum_mapping(x: Mapping, **kwargs) -> int:
 
 
 def _terminate_checksum(x: int, fullname: str, **kwargs) -> int:
+    """Returns checksum for final value + name + accumulator."""
     return x + _cached_checksum_str(fullname) + kwargs.get("accumulator", 0)
+
+
+def _add_type_checksum_to_accumulator(x: Any, kwargs: Dict[str, Any]) -> Dict[str, Any]:
+    kwargs["accumulator"] = kwargs.get("accumulator", 0) + _cached_checksum_str(
+        get_fullname(x)
+    )
+    return kwargs
 
 
 @lru_cache(maxsize=None)